Don't overwrite /etc/default/qemu-kvm profile.
The previous change Ib91ca930a91da3d75dbd96f7d55e7e259cb50fd1
enforces the replacement of the system's file /etc/default/qemu-kvm with
a template that doesn't reflects all the possible configuration
values available for qemu-kvm to work. For instance, LP: #1670281 raises the
issue for the KSM configuration.

This patch removes the template, instead it does
a sed replacement over the file to switch the KVM_HUGEPAGES
setting on/off, preventing to replace the system's configured values.

Change-Id: I523587720bfea14150e4d3de7b551bbe4cd28c66
Closes-Bug: #1670281